Move `util_*` files used only by `io` to `io.conversions`.
